requirements.txt
pyspark==3.5.1
kafka-python
fastapi
uvicorn
pymongo
streamlit
python-dotenv
pydantic
